Create New Patient page with one control
